third_party/blink/web_tests/wpt_internal/app-history/navigate-event/respondWith-reject.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  appHistory.onnavigate = e => {
    e.respondWith(new Promise((resolve, reject) => t.step_timeout(() => {
      reject();
      t.step_timeout(t.step_func_done(() => assert_equals(location.hash, "#1")), 0);
    }, 0)));
  };

  location.href = "#1";
  assert_equals(location.hash, "#1");
}, "AppHistoryNavigateEvent.respondWith should abort if the given promise rejects");
</script>
